removed outdated asynchronous tests, fixed the redirect streaming bug
